+++ app/jobs/garbage_collection_job.rb
@@ -9,6 +9,6 @@ class GarbageCollectionJob < ApplicationJob
       return
     end
 
-    GarbageCollectionRunner.call
+    GraphIntegrityService.call
   end
 end